libant: Add ant_is_dead, mark dead upon read failures
If a read fails in an unrecoverable way (ie. not a timeout) then the ant
transport is dead. Keep track of this, so that users like libfitbit can
check at opportune moments whether the transport died and destroy the
ant instance. Perform this check in fitbit_sync_trackers, such that a
failed sync due to a dead transport returns an error as it should.