fitbitd: Modularise the install target
Rather than the top-level makefile knowing too many details about
fitbitd, specify a general install target and an install-fitbitd target
that it depends upon. This will also allow other install targets to be
added soon (eg. for an indicator) and allow each to be targeted
individually (eg. for a split package) or all together.